requirements.txt
chardet~=5.1.0
Django~=4.2
jieba~=0.42.1
jieba_fast~=0.53
pandas~=2.0.1
python-dotenv~=1.0.0
mysqlclient~=2.1.1
django-redis~=5.2.0
openpyxl~=3.1.2
xlwt~=1.3.0
beautifulsoup4~=4.12.2
selenium~=4.9.0
concurrent-log-handler~=0.9.19
celery~=5.2.7
gevent~=22.10.2